More WIP for AO

+++ com.unity.render-pipelines.universal/Shaders/Utils/ScreenSpaceAmbientOcclusion.shader
@@ -20,8 +20,8 @@ Shader "Hidden/Universal Render Pipeline/ScreenSpaceAmbientOcclusion"
 
         struct Varyings
         {
-            half4  positionCS   : SV_POSITION;
-            half4  uv           : TEXCOORD0;
+            float4  positionCS   : SV_POSITION;
+            float4  uv           : TEXCOORD0;
             UNITY_VERTEX_OUTPUT_STEREO
         };
